refactor: use updated code style
