+++ src/app/about/about.component.ts
@@ -29,24 +29,24 @@ export class About {
   ngOnInit() {
     console.log('hello `About` component');
     // static data that is bundled
-    var mockData = require('assets/mock-data/mock-data.json');
-    console.log('mockData', mockData);
+    // var mockData = require('assets/mock-data/mock-data.json');
+    // console.log('mockData', mockData);
     // if you're working with mock data you can also use http.get('assets/mock-data/mock-data.json')
-    this.asyncDataWithWebpack();
+    // this.asyncDataWithWebpack();
   }
   asyncDataWithWebpack() {
     // you can also async load mock data with 'es6-promise-loader'
     // you would do this if you don't want the mock-data bundled
     // remember that 'es6-promise-loader' is a promise
-    var asyncMockDataPromiseFactory = require('es6-promise!assets/mock-data/mock-data.json');
-    setTimeout(() => {
-
-      let asyncDataPromise = asyncMockDataPromiseFactory();
-      asyncDataPromise.then(json => {
-        console.log('async mockData', json);
-      });
-
-    });
+    // var asyncMockDataPromiseFactory = require('es6-promise!assets/mock-data/mock-data.json');
+    // setTimeout(() => {
+    //
+    //   let asyncDataPromise = asyncMockDataPromiseFactory();
+    //   asyncDataPromise.then(json => {
+    //     console.log('async mockData', json);
+    //   });
+    //
+    // });
   }
 
 }
