+++ src/app/about/about.spec.ts
@@ -10,7 +10,7 @@ import {
 import {Component, provide} from 'angular2/core';
 
 // Load the implementations that should be tested
-import {About} from './about';
+import {About} from './about.component';
 
 describe('About', () => {
   // provide our implementations or mocks to the dependency injector
