+++ src/app/about/index.ts
@@ -1,1 +1,1 @@
-export * from './about';
+export * from './about.component';
