+++ src/app/app.component.ts
@@ -6,6 +6,7 @@ import {RouteConfig, Router} from 'angular2/router';
 
 import {Home} from './home';
 import {AppState} from './app.service';
+import {RouterActive} from './router-active';
 
 /*
  * App Component
@@ -15,7 +16,7 @@ import {AppState} from './app.service';
   selector: 'app',
   pipes: [ ],
   providers: [ ],
-  directives: [ ],
+  directives: [ RouterActive ],
   styles: [`
     h1 {
       font-family: Arial, Helvetica, sans-serif
