+++ src/app/home/home.spec.ts
@@ -13,8 +13,8 @@ import {MockBackend} from 'angular2/http/testing';
 import {WebpackState} from 'angular2-hmr';
 
 // Load the implementations that should be tested
-import {Home} from './home';
-import {Title} from './services/title';
+import {Home} from './home.component';
+import {Title} from './title';
 import {AppState} from '../app.service';
 
 describe('Home', () => {
