+++ src/app/home/index.ts
@@ -1,1 +1,1 @@
-export * from './home';
+export * from './home.component';
