+++ src/app/home/title/index.ts
@@ -1,0 +1,1 @@
+export * from './title.service';
