+++ src/app/home/x-large/index.ts
@@ -1,0 +1,1 @@
+export * from './x-large.directive';
