+++ src/app/home/x-large/x-large.spec.ts
@@ -12,7 +12,7 @@ import {BaseRequestOptions, Http} from 'angular2/http';
 import {MockBackend} from 'angular2/http/testing';
 
 // Load the implementations that should be tested
-import {XLarge} from './x-large';
+import {XLarge} from './x-large.directive';
 
 describe('x-large directive', () => {
   // Create a test component to test directives
