+++ src/app/index.ts
@@ -1,2 +1,3 @@
 // App
-export * from './app';
+export * from './app.component';
+export * from './app.service';
