+++ src/app/router-active/index.ts
@@ -1,0 +1,2 @@
+// Application level directive
+export * from './router-active.directive';
