+++ src/main.browser.ts
@@ -14,8 +14,7 @@ import {ENV_PROVIDERS} from './platform/environment';
 * App Component
 * our top level component that holds all of our components
 */
-import {AppState} from './app/app.service';
-import {App} from './app/app';
+import {App, AppState} from './app';
 
 /*
  * Bootstrap our Angular app with a top level component `App` and inject
