--- src/platform/browser/router-active/index.ts
@@ -1,2 +1,0 @@
-// Application level directive
-export * from './router-active';
